#211: Remove duplicated file, renamed ractiveConcatComponents to ractiveConcatObjects
