+++ tasks/ractiveConcatObjects.js
